# openapi.yaml
openapi: 3.0.3
info:
  title: Go Kasir API
  version: 1.0.0
  description: Simple cashier API for product CRUD, checkout, and sales reporting.
servers:
  - url: /
paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      responses:
        "200":
          description: API is running
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - message
                properties:
                  status:
                    type: string
                    example: OK
                  message:
                    type: string
                    example: API Running
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  user:
                    type: object
        "400":
          description: Invalid request
        "401":
          description: Invalid credentials
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Logout and clear session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /api/products:
    get:
      tags:
        - Products
      summary: Get all products
      security:
        - SessionCookie: []
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filter products by name (case-insensitive)
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
    post:
      tags:
        - Products
      summary: Create a product
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
  /api/products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product ID
        "404":
          description: Product not found
        "401":
          description: Unauthorized
    put:
      tags:
        - Products
      summary: Update product by ID
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request
        "404":
          description: Product not found
        "401":
          description: Unauthorized
    delete:
      tags:
        - Products
      summary: Delete product by ID
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Invalid product ID
        "404":
          description: Product not found
        "401":
          description: Unauthorized
  /api/categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      security:
        - SessionCookie: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
    post:
      tags:
        - Categories
      summary: Create a category
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
  /api/categories/{id}:
    get:
      tags:
        - Categories
      summary: Get category by ID
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid category ID
        "404":
          description: Category not found
        "401":
          description: Unauthorized
    put:
      tags:
        - Categories
      summary: Update category by ID
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid request
        "404":
          description: Category not found
        "401":
          description: Unauthorized
    delete:
      tags:
        - Categories
      summary: Delete category by ID
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Invalid category ID
        "404":
          description: Category not found
        "401":
          description: Unauthorized
  /api/checkout:
    post:
      tags:
        - Transactions
      summary: Create a checkout transaction
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
  /api/report/hari-ini:
    get:
      tags:
        - Reports
      summary: Sales summary for today
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesSummary"
        "401":
          description: Unauthorized
  /api/report:
    get:
      tags:
        - Reports
      summary: Sales summary by date range
      security:
        - SessionCookie: []
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesSummary"
        "400":
          description: Missing or invalid dates
        "401":
          description: Unauthorized
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
        created_at:
          type: string
          format: date-time
    Product:
      type: object
      required:
        - id
        - name
        - price
        - stock
        - category_id
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Instant Noodles
        price:
          type: integer
          example: 3500
        stock:
          type: integer
          example: 10
        category_id:
          type: integer
          example: 2
        category_name:
          type: string
          example: Beverages
    ProductInput:
      type: object
      required:
        - name
        - price
        - stock
        - category_id
      properties:
        name:
          type: string
          example: Coffee
        price:
          type: integer
          example: 2500
        stock:
          type: integer
          example: 200
        category_id:
          type: integer
          example: 2
    Category:
      type: object
      required:
        - id
        - name
        - description
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Beverages
        description:
          type: string
          example: Drinks and bottled water
    CategoryInput:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          example: Snacks
        description:
          type: string
          example: Chips, biscuits, and light bites
    Message:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: delete success
    Transaction:
      type: object
      required:
        - id
        - total_amount
        - created_at
        - details
      properties:
        id:
          type: integer
          example: 1
        total_amount:
          type: integer
          example: 45000
        created_at:
          type: string
          format: date-time
        details:
          type: array
          items:
            $ref: "#/components/schemas/TransactionDetail"
    TransactionDetail:
      type: object
      required:
        - id
        - transaction_id
        - product_id
        - quantity
        - subtotal
      properties:
        id:
          type: integer
          example: 1
        transaction_id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 2
        product_name:
          type: string
          example: Instant Noodles
        quantity:
          type: integer
          example: 2
        subtotal:
          type: integer
          example: 7000
    CheckoutItem:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
          example: 2
        quantity:
          type: integer
          example: 3
    CheckoutRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CheckoutItem"
    SalesSummary:
      type: object
      required:
        - total_revenue
        - total_transactions
      properties:
        total_revenue:
          type: integer
          example: 45000
        total_transactions:
          type: integer
          example: 5
        top_product:
          $ref: "#/components/schemas/TopProduct"
    TopProduct:
      type: object
      required:
        - name
        - qty_sold
      properties:
        name:
          type: string
          example: Instant Noodles
        qty_sold:
          type: integer
          example: 12